FixTest

+++ src/main/java/daos/core/AlertDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import entities.core.Alert;
 
-public interface AlertDao extends JpaRepository<Alert, Long> {
+public interface AlertDao extends JpaRepository<Alert, Integer> {
 
 }
